requirements.txt
# Core
pandas>=2.0.0
numpy>=1.23.0

# Machine Learning
scikit-learn>=1.6.0
xgboost>=2.0.0
joblib>=1.3.0

# Dashboard
streamlit>=1.36.0

# Visual Features
pillow>=10.0.0
opencv-python>=4.9.0
imagehash>=4.3.1
colormath>=3.0.0
tldextract>=5.1.0
requests>=2.31.0

# OCR (EasyOCR replaces pytesseract)
easyocr>=1.7.1

# Browser Automation
playwright>=1.42.0

# GeoIP
geoip2>=4.7.0

# Vector Search (Milvus + embeddings)
pymilvus>=2.4.0
sentence-transformers>=2.7.0
torch>=2.2.0
torchvision
transformers>=4.41.0
tf-keras  # ✅ required to fix the Keras 3 incompatibility
fpdf
jellyfish
rapidfuzz
python-whois
# Excel
openpyxl>=3.1.2

# Misc
python-magic>=0.4.27
dnspython>=2.6.0
